Volumes adapter + refactor
Volumes and Jobs both use the composite id pattern of
ID + Namespace. This refactors the logic for this from
the Job adapter into a mixin.
