+++ ui/app/adapters/volume.js
@@ -1,0 +1,4 @@
+import Watchable from './watchable';
+import WithNamespaceIDs from 'nomad-ui/mixins/with-namespace-ids';
+
+export default Watchable.extend(WithNamespaceIDs);
